requirements.txt
Flask==2.2.3
Flask-Session==0.4.0
werkzeug==2.2.3
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=1.0.1
gunicorn==20.1.0
aiofiles<24.0,>=22.0

pydantic~=2.10.4
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
numpy
#datasets~=3.2.0
#fastapi~=0.115.11

#html2text~=2024.2.26
#gymnasium~=1.0.0
#pillow~=10.4.0
#browsergym~=0.13.3
#uvicorn~=0.34.0
#unidiff~=0.7.5
#browser-use~=0.1.40
#googlesearch-python~=1.3.0
#baidusearch~=1.0.3
#duckduckgo_search~=7.5.1

#pydantic_core~=2.27.2
colorama~=0.4.6
#playwright~=1.49.1
